feat(landing): added new Landing page to replace Login page, removed extraneous elements from Login
