+++ src/pages/landing/landing.css
@@ -1,0 +1,14 @@
+
+.landing a {
+  color: #aaa;
+}
+.landing a:not(:hover) {
+  text-decoration: none;
+}
+.landing a.emphasis {
+  color: #ccb0ff;
+}
+.landing small {
+  font-size: calc(6px + 0.75vmin);
+  color: rgba(255, 255, 255, 0.42);
+}
